+++ packages/node/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add the sorabanTxs field to transaction information(#112)
+
 ## [4.6.0] - 2025-03-05
 ### Changed
 - Support http soroban endpoints (#109)